// angular/src/app/api/model/clienteJuridicoListResponse.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Dados resumidos de cliente pessoa jurídica para listas
 */
export interface ClienteJuridicoListResponse {
  /**
   * ID do cliente
   */
  id?: number;
  /**
   * Razão Social
   */
  razaoSocial?: string;
  /**
   * CNPJ
   */
  cnpj?: string;
  /**
   * E-mail do cliente
   */
  email?: string;
  /**
   * Status do cliente
   */
  estaAtivo?: boolean;
}